% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svggage.R
\name{svg_gauge}
\alias{svg_gauge}
\title{Svg Guage}
\usage{
svg_gauge(value, max, showValue = TRUE, width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{value}{Numeric value of the guage}

\item{max}{Maximium value in the gauge}

\item{showValue}{If TRUE show the value in the middle of the guage.}

\item{width}{Width of the htmlwidget}

\item{height}{Height of the htmlwidget}

\item{elementId}{ElementID to reference}
}
\description{
Generates a an SVG guage
}
